requirements.txt
wheel==0.23.0
hexdump==3.2
bitarray==0.8.1
PyYAML==3.10
pymongo==3.0.3
